devsite: add remaining files from original site
We previously hosted all of the files from the original site when we had
them in a compiled form; we add the source for these now, since that's
about the same (the only difference is that we have the source to
compile from!).

All of these files are imported from the original dev site, and there is
no new content here.

Signed-off-by: Eric Migicovsky <eric@repebble.com>