Update hardware information to include new Core Devices watches
Some of the information is incomplete at this stage, since it is still not confirmed. I could not find any information as to what CPU the Pebble 2 Duo is. I suspect it is the same as the Pebble 2, but am not certain. If I get more information in future I will make another PR to fill out the missing data.